+++ src/services/content/content.js
@@ -1,5 +1,5 @@
 import { useQuery } from 'react-query';
-import { getRequest } from '../utils/requests';
+import { getRequest } from '../../utils/requests';
 import { UATCMSURL } from './content.constants';
 
 const baseURL = require('../pages/commonComponents/BaseURL');
@@ -7,8 +7,8 @@ const cmsURL = baseURL.getCMSURL();
 const env = baseURL.getURLEnv();
 const local = 'local';
 
-export const getNon5SafesModalContentRequest = (options) => {
-    const url = env === local ? UATCMSURL : cmsURL;
+export const getNon5SafesModalContentRequest = options => {
+	const url = env === local ? UATCMSURL : cmsURL;
 	return getRequest(`${url}/Non5SafesModalContent`, options);
 };
 
@@ -20,5 +20,5 @@ const useGetNon5SafesModalContentRequest = (requestOptions, queryOptions = { que
 };
 
 export default {
-    useGetNon5SafesModalContentRequest
-}
+	useGetNon5SafesModalContentRequest,
+};
