+++ src/services/kpis/kpis.js
@@ -1,6 +1,6 @@
 import { useQuery } from 'react-query';
-import { apiURL } from '../configs/url.config';
-import { getRequest } from '../utils/requests';
+import { apiURL } from '../../configs/url.config';
+import { getRequest } from '../../utils/requests';
 
 const getKpis = options => {
 	return getRequest(`${apiURL}/kpis`, options);
