+++ src/services/kpis/kpis.test.js
@@ -1,11 +1,11 @@
 import { renderHook } from '@testing-library/react-hooks';
 import { QueryClient, QueryClientProvider } from 'react-query';
-import { apiURL } from '../configs/url.config';
-import { getRequest } from '../utils/requests';
+import { apiURL } from '../../configs/url.config';
+import { getRequest } from '../../utils/requests';
 import service from './kpis';
 
 jest.mock('axios');
-jest.mock('../utils/requests');
+jest.mock('../../utils/requests');
 
 let wrapper;
