+++ src/services/reviews/index.js
@@ -1,0 +1,1 @@
+export { default } from './reviews';
